* complete refactor of unarchiver.py and move to plugin.py
